Many UI fixes have applied when poll posts where consecutive.

+++ messaging/GLPGroupManager.h
@@ -14,7 +14,7 @@
 
 + (NSDictionary *)processGroups:(NSArray *)groups;
 
-+ (void)loadInitialPostsWithGroupId:(int)groupId localCallback:(void (^)(NSArray *localPosts))localCallback remoteCallback:(void (^)(BOOL success, BOOL remain, NSArray *remotePosts))remoteCallback;
++ (void)loadInitialPostsWithGroupId:(NSInteger)groupId localCallback:(void (^)(NSArray *localPosts))localCallback remoteCallback:(void (^)(BOOL success, BOOL remain, NSArray *remotePosts))remoteCallback;
 
 + (void)loadGroups:(NSArray *)groups withLocalCallback:(void (^)(NSArray *groups))localCallback remoteCallback:(void (^)(BOOL success, NSArray *groups))remoteCallback;
